Move extractQuery into url, add params to url
