+++ packages/jaeger-ui/src/components/DeepDependencyGraph/url.test.js
@@ -12,9 +12,10 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+import queryString from 'query-string';
 import * as reactRouterDom from 'react-router-dom';
 
-import { ROUTE_PATH, matches, getUrl } from './url';
+import { ROUTE_PATH, matches, getUrl, getUrlState } from './url';
 
 describe('DeepDependencyGraph/url', () => {
   describe('matches', () => {
@@ -43,8 +44,104 @@ describe('DeepDependencyGraph/url', () => {
   });
 
   describe('getUrl', () => {
-    it('handles an empty state', () => {
+    it('handles no args given', () => {
+      expect(getUrl()).toBe('/deep-dependencies');
+    });
+
+    it('handles empty args', () => {
       expect(getUrl({})).toBe('/deep-dependencies');
+    });
+
+    it('includes provided args', () => {
+      const paramA = 'aParam';
+      const paramB = 'bParam';
+      expect(getUrl({ paramA, paramB })).toBe(`/deep-dependencies?paramA=${paramA}&paramB=${paramB}`);
+    });
+  });
+
+  describe('getUrlState', () => {
+    const search = 'test search';
+    const service = 'serviceName';
+    const operation = 'operationName';
+    const start = '400';
+    const end = '900';
+    const acceptableParams = {
+      service,
+      operation,
+      start,
+      end,
+    };
+    const expectedParams = {
+      service,
+      operation,
+      start: Number.parseInt(start, 10),
+      end: Number.parseInt(end, 10),
+    };
+    let warnSpy;
+    let parseSpy;
+
+    beforeAll(() => {
+      parseSpy = jest.spyOn(queryString, 'parse');
+      warnSpy = jest.spyOn(console, 'warn').mockImplementation();
+    });
+
+    beforeEach(() => {
+      parseSpy.mockReset();
+      warnSpy.mockReset();
+    });
+
+    afterAll(() => {
+      warnSpy.mockRestore();
+    });
+
+    it('gets all values from queryString', () => {
+      parseSpy.mockReturnValue(acceptableParams);
+      expect(getUrlState(search)).toEqual(expectedParams);
+      expect(parseSpy).toHaveBeenCalledWith(search);
+    });
+
+    it('handles absent values', () => {
+      ['service', 'operation', 'start', 'end'].forEach(param => {
+        const { [param]: unused, ...rest } = expectedParams;
+        parseSpy.mockReturnValue(rest);
+        expect(getUrlState(search)).toEqual(rest);
+        expect(parseSpy).toHaveBeenLastCalledWith(search);
+      });
+    });
+
+    it('ignores extraneous query parameters', () => {
+      const extraneous = {
+        param: 'value',
+      };
+      parseSpy.mockReturnValue({
+        ...extraneous,
+        ...acceptableParams,
+      });
+      expect(getUrlState(search)).toEqual(expect.not.objectContaining(extraneous));
+      expect(parseSpy).toHaveBeenCalledWith(search);
+    });
+
+    it('omits falsy values', () => {
+      ['service', 'operation', 'start', 'end'].forEach(param => {
+        [null, undefined, ''].forEach(falsyPossibility => {
+          parseSpy.mockReturnValue({ ...expectedParams, [param]: falsyPossibility });
+          expect(Reflect.has(getUrlState(search), param)).toBe(false);
+          expect(parseSpy).toHaveBeenLastCalledWith(search);
+        });
+      });
+    });
+
+    it('handles and warns on duplicate values', () => {
+      ['service', 'operation', 'start', 'end'].forEach(param => {
+        const secondParam = `second ${acceptableParams[param]}`;
+        parseSpy.mockReturnValue({
+          ...acceptableParams,
+          [param]: [acceptableParams[param], secondParam],
+        });
+        expect(getUrlState(search)[param]).toBe(expectedParams[param]);
+        expect(warnSpy).toHaveBeenLastCalledWith(expect.stringContaining(secondParam));
+        expect(parseSpy).toHaveBeenLastCalledWith(search);
+      });
     });
   });
 });
